oceanstate_analysis/requirements.txt
-e.
black
flake8
isort
loguru
pip
python-dotenv
tqdm
typer
pathlib
pandas
matplotlib
seaborn
ipykernel
pandas